+++ guidedmodules/app_loading.py
@@ -1,5 +1,5 @@
 ###########################################################
-# Import Apps into the AppInstance, Module, ModuleQuestion,
+# Import Apps into the AppVersion, Module, ModuleQuestion,
 # and ModuleAsset Django ORM models.
 ###########################################################
 
@@ -11,7 +11,7 @@ from collections import OrderedDict
 from django.db import transaction
 from django.db.models.deletion import ProtectedError
 
-from .models import AppSource, AppInstance, ModuleAsset, \
+from .models import AppSource, AppVersion, ModuleAsset, \
                     Module, ModuleQuestion, Task, \
                     extract_catalog_metadata
 
@@ -49,9 +49,9 @@ def load_app_into_database(app, update_mode=AppImportUpdateMode.CreateInstance, 
     # be processed recursively.
     available_modules = dict(app.get_modules())
 
-    # Create an AppInstance to add new Modules into, unless update_appinst is given.
+    # Create an AppVersion to add new Modules into, unless update_appinst is given.
     if update_appinst is None:
-        appinst = AppInstance.objects.create(
+        appinst = AppVersion.objects.create(
             source=app.store.source,
             appname=app.name,
             catalog_metadata={},
@@ -76,11 +76,31 @@ def load_app_into_database(app, update_mode=AppImportUpdateMode.CreateInstance, 
     load_module_assets_into_database(app, appinst)
 
     # If there's an 'app' module, move the app catalog information
-    # to the AppInstance.
+    # to the AppVersion.
     if 'app' in processed_modules:
         extract_catalog_metadata(processed_modules['app'])
-        appinst.save()
         processed_modules['app'].save()
+
+    # If there's a README.md file, overwrite the app catalog description.
+    import re, fs.errors
+    try:
+        # Read the README.md.
+        readme = app.read_file("README.md")
+
+        # Strip any initial heading that has the app name itself, since
+        # that is expected to not be included in the long description.
+        # Check both CommonMark heading formats.
+        readme = re.sub(r"^\s*#+ *" + re.escape(appinst.catalog_metadata["title"]) + r"\s*", "", readme)
+        readme = re.sub(r"^\s*" + re.escape(appinst.catalog_metadata["title"]) + r"\s*[-=]+\s*", "", readme)
+
+        appinst.catalog_metadata\
+            .setdefault("description", {})["long"] = readme
+    except fs.errors.ResourceNotFound:
+        pass
+
+    # Update appinst. It may have been modified by extract_catalog_metadata
+    # and by the loading of a README.md file.
+    appinst.save()
 
     return appinst