+++ siteapp/tests.py
@@ -25,12 +25,13 @@ class SeleniumTest(StaticLiveServerTestCase):
         from django.conf import settings
         settings.EMAIL_BACKEND = 'django.core.mail.backends.locmem.EmailBackend'
 
-        # Override ALLOWED_HOSTS, SITE_ROOT_URL, and ORGANIZATION_PARENT_DOMAIN
+        # Override ALLOWED_HOSTS, SITE_ROOT_URL, ORGANIZATION_PARENT_DOMAIN, etc.
         # because they may not be set or set properly in the local environment's
         # non-test settings for the URL assigned by the LiveServerTestCase server.
         settings.ALLOWED_HOSTS = ['.localhost', 'testserver']
         settings.SITE_ROOT_URL = cls.live_server_url
         settings.ORGANIZATION_PARENT_DOMAIN = 'orgs.localhost'
+        settings.SINGLE_ORGANIZATION_KEY = None
 
         # In order for these tests to succeed when not connected to the
         # Internet, disable email deliverability checks which query DNS.
@@ -151,6 +152,7 @@ class OrganizationSiteFunctionalTests(SeleniumTest):
         # Load the Q modules from the fixtures directory.
         from guidedmodules.models import AppSource
         from guidedmodules.management.commands.load_modules import Command as load_modules
+        
         AppSource.objects.all().delete()
         AppSource.objects.get_or_create(
               # this one exists on first db load because it's created by
@@ -165,6 +167,8 @@ class OrganizationSiteFunctionalTests(SeleniumTest):
                 }
             }
         )
+        load_modules().handle() # load system modules
+
         AppSource.objects.create(
             slug="project",
             spec={ # contains a test project
@@ -172,8 +176,8 @@ class OrganizationSiteFunctionalTests(SeleniumTest):
                 "path": "fixtures/modules/other",
             },
             trust_assets=True
-        )
-        load_modules().handle()
+        )\
+            .add_app_to_catalog("simple_project")
 
         # Create a default user that is a member of the organization.
         # Log the user into the test client, which is used for API
@@ -244,13 +248,13 @@ class OrganizationSiteFunctionalTests(SeleniumTest):
         self.fill_field("#id_password", password or self.user.clear_password)
         self.click_element("form button.primaryAction")
 
-    def _new_project(self, module_key="project/simple_project"):
+    def _new_project(self):
         self.browser.get(self.url("/projects"))
         self.click_element("#new-project")
-        self.click_element(".app[data-app='%s'] .view-app" % module_key)
+        self.click_element(".app[data-app='project/simple_project'] .view-app")
         self.click_element("#start-project")
         # last two lines could also be replaced with:
-        #self.click_element(".app[data-app='%s'] .start-app" % module_key)
+        #self.click_element(".app[data-app='project/simple_project'] .start-app")
         var_sleep(1)
         self.assertRegex(self.browser.title, "I want to answer some questions on Q.")